Moving functions out of SudokuBoard.tsx
