+++ sudokuru/app/Components/SudokuBoard/Core/Components/Puzzle.tsx
@@ -1,27 +1,33 @@
 import { View } from "react-native";
 import React from "react";
-import {
-  CellProps,
-  SudokuObjectProps,
-} from "../../../../Functions/LocalDatabase";
+import { SudokuObjectProps } from "../../../../Functions/LocalDatabase";
+import { HintObjectProps } from "../../SudokuBoard";
 
 interface PuzzleProps {
-  renderCell: (
-    cell: CellProps,
+  RenderCell: (
+    sudokuBoard: SudokuObjectProps,
+    setSudokuBoard: React.Dispatch<
+      React.SetStateAction<SudokuObjectProps | undefined>
+    >,
+    sudokuHint: HintObjectProps | undefined,
     r: number,
     c: number
   ) => JSX.Element | undefined;
   sudokuBoard: SudokuObjectProps;
+  setSudokuBoard: React.Dispatch<
+    React.SetStateAction<SudokuObjectProps | undefined>
+  >;
+  sudokuHint: HintObjectProps | undefined;
 }
 
 const Puzzle = (props: PuzzleProps) => {
-  const { sudokuBoard, renderCell } = props;
+  const { RenderCell, sudokuBoard, setSudokuBoard, sudokuHint } = props;
 
   const renderAllRows = [];
   for (let r = 0; r < 9; r++) {
     const rows = [];
     for (let c = 0; c < 9; c++) {
-      rows.push(renderCell(sudokuBoard.puzzle[r][c], r, c));
+      rows.push(RenderCell(sudokuBoard, setSudokuBoard, sudokuHint, r, c));
     }
     renderAllRows.push(rows);
   }
